+++ src/grafzahl/description_printer.rs
@@ -1,4 +1,4 @@
-use crate::grafv4::countable::Countable;
+use crate::grafzahl::countable::Countable;
 
 pub(crate) fn print_description<CountMode: Countable>() {
     CountMode::display_description();
