+++ src/grafzahl/ignore_checker.rs
@@ -7,6 +7,8 @@ use crate::{Cli, get_config_location, AppState};
 
 pub fn init_ignore_list() -> Gitignore {
     //TODO ignore_list should have file extension .gitignore
+    //TODO should take path from some where else
+    //TODO should be checked if path is overridden with env
     let path_buff = PathBuf::from(format!("{}/ignore_list.txt", get_config_location()));
     let args = Cli::from_args();
     let mut builder = GitignoreBuilder::new(&args.directory);
