+++ src/grafzahl/mod.rs
@@ -1,2 +1,8 @@
 pub mod languages;
 pub mod ignore_checker;
+pub mod count_mode;
+pub mod description_printer;
+mod generic_counter;
+mod io_reader;
+mod countable;
+mod count_mode_lib;
